ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.3
scikit-learn==1.4.0
sentence-transformers==2.2.2
qdrant-client==1.7.0
redis==5.0.1
python-multipart==0.0.6
